requirements.txt
# Core Dependencies
mistralai==0.0.11
boto3==1.26.137
chromadb==0.3.21
python-dotenv==1.0.0

# Web Framework
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5

# Data Processing
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Database
sqlalchemy==2.0.19

# Scheduling
apscheduler==3.10.4

# Machine Learning (Optional)
scikit-learn==1.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
